Router files migrated to typescript
